pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wandb-generic"
version = "1.0.0"
authors = [
    { name="WandB Generic Contributors" },
]
description = "A truly generic Python package for Weights & Biases integration with any ML/DL library or long-running function."
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
keywords = [
    "wandb", "weights-and-biases", "machine-learning", "deep-learning", 
    "experiment-tracking", "logging", "pytorch", "tensorflow", "generic"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Logging",
    "Typing :: Typed",
]
dependencies = [
    "wandb>=0.15.0",
    "pyyaml>=6.0",
    "numpy>=1.19.0",
]

[project.optional-dependencies]
torch = ["torch>=1.9.0"]
tensorflow = ["tensorflow>=2.6.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "isort>=5.10",
    "mypy>=1.0",
    "pre-commit>=2.20",
]
all = [
    "torch>=1.9.0",
    "tensorflow>=2.6.0",
]

[project.urls]
Homepage = "https://github.com/thedatageek/wandb-generic"
Documentation = "https://github.com/thedatageek/wandb-generic#readme"
Repository = "https://github.com/thedatageek/wandb-generic"
"Bug Tracker" = "https://github.com/thedatageek/wandb-generic/issues"
Changelog = "https://github.com/thedatageek/wandb-generic/releases"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
wandb_generic = ["py.typed"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true